WiFi connection manager for sensors and actuers
ConnectionManager:
Provide an standardised connection to the WiFi network including saving
the network´s data in the EEPROM and building an configuration page for
the user.
The access to this system is grandet by the function begin(), which
takes care about everything and stop the script  from running until a
connection is successfull.

DNSServer:
From https://www.hackster.io/rayburne/esp8266-captive-portal-5798ff
under MIT license

Used to rewrite all requests over the ESP´s AP to the configuration page